Extract TablesGrid to own file
